Refactor structure for ombud components.
