+++ src/App.vue
@@ -53,7 +53,7 @@ import TheNavigationDrawer from "@/components/TheNavigationDrawer.vue";
 import TheSnackbar from "@/components/TheSnackbar.vue";
 import TheFooter from "@/components/TheFooter.vue";
 import FeedbackMenu from "@/components/feedback/FeedbackMenu.vue";
-import OmbudsMenu from "@/components/OmbudsMenu.vue";
+import OmbudsMenu from "@/components/ombud/OmbudsMenu.vue";
 
 import { log } from "@/utils/log";
